+++ clouds/databricks/modules/dummy/sql/dummy_function.sql
@@ -1,0 +1,2 @@
+-- Dummy Function
+CREATE OR REPLACE FUNCTION dummy_function as 'com.carto.analyticstoolbox.modules.dummy.DummyFunction';
